requirements.txt
Flask
flask-cors
Pillow
arabic-reshaper
python-bidi
requests
